+++ trace/opentracing.go
@@ -20,9 +20,10 @@ import (
 
 // Lists the names of headers that a specification uses for representing trace information.
 type HeaderGroup struct {
-	TraceID   string
-	SpanID    string
-	numFormat headerGroupNumFormat
+	TraceID         string
+	SpanID          string
+	numFormat       headerGroupNumFormat
+	OutgoingHeaders map[string]string
 }
 
 type headerGroupNumFormat int
@@ -42,6 +43,9 @@ var HeaderFormats = []HeaderGroup{
 		TraceID:   "ot-tracer-traceid",
 		SpanID:    "ot-tracer-spanid",
 		numFormat: hexadecimal,
+		OutgoingHeaders: map[string]string{
+			"ot-tracer-sampled": "true",
+		},
 	},
 	// OpenTracing format.
 	HeaderGroup{
@@ -59,6 +63,10 @@ var HeaderFormats = []HeaderGroup{
 		SpanID:  "Spanid",
 	},
 }
+
+// defaultHeaderFormat is the way .Inject sets HTTP headers by
+// default.
+var defaultHeaderFormat = HeaderFormats[0]
 
 // GlobalTracer is the… global tracer!
 var GlobalTracer = Tracer{}
@@ -514,7 +522,6 @@ func (tracer Tracer) ExtractRequestChild(resource string, req *http.Request, nam
 // Inject injects the provided SpanContext into the carrier for propagation.
 // It will return opentracing.ErrUnsupportedFormat if the format is not supported.
 // TODO support other SpanContext implementations
-// TODO support all the BuiltinFormats
 func (t Tracer) Inject(sm opentracing.SpanContext, format interface{}, carrier interface{}) (err error) {
 	defer func() {
 		if r := recover(); r != nil {
@@ -528,7 +535,8 @@ func (t Tracer) Inject(sm opentracing.SpanContext, format interface{}, carrier i
 		return ErrUnsupportedSpanContext
 	}
 
-	if format == opentracing.Binary {
+	switch format {
+	case opentracing.Binary:
 		// carrier is guaranteed to be an io.Writer by contract
 		w := carrier.(io.Writer)
 
@@ -541,6 +549,20 @@ func (t Tracer) Inject(sm opentracing.SpanContext, format interface{}, carrier i
 		}
 
 		return trace.ProtoMarshalTo(w)
+	case opentracing.HTTPHeaders:
+		h := carrier.(opentracing.HTTPHeadersCarrier)
+		base := 10
+		if defaultHeaderFormat.numFormat == hexadecimal {
+			base = 16
+		}
+		h.Set(defaultHeaderFormat.SpanID, strconv.FormatInt(sc.SpanID(), base))
+		h.Set(defaultHeaderFormat.TraceID, strconv.FormatInt(sc.TraceID(), base))
+		if defaultHeaderFormat.OutgoingHeaders != nil {
+			for name, value := range defaultHeaderFormat.OutgoingHeaders {
+				h.Set(name, value)
+			}
+		}
+		return nil
 	}
 
 	// If the carrier is a TextMapWriter, treat it as one, regardless of what the format is